requirements.txt
# Key Inventory Management System Requirements
# Python dependencies for the enhanced key lifecycle management system

# Core dependencies
pyyaml>=6.0.1
jsonschema>=4.17.0
requests>=2.28.0

# Cloud provider SDKs for key store integration
boto3>=1.26.0          # AWS KMS integration
botocore>=1.29.0       # AWS core library

# Azure Key Vault integration
azure-identity>=1.12.0
azure-keyvault-keys>=4.8.0
azure-keyvault-secrets>=4.7.0

# HashiCorp Vault integration
hvac>=1.0.2

# Development and testing
pytest>=7.2.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional: Enhanced logging and monitoring
structlog>=22.0.0      # Structured logging
prometheus-client>=0.15.0  # Metrics collection

# Optional: Security scanning
safety>=2.3.0          # Security vulnerability scanning
bandit>=1.7.0          # Security linting

# Optional: Documentation generation
mkdocs>=1.4.0
mkdocs-material>=8.5.0

# Legacy compatibility
pydantic[email]>=1.10.0
click>=8.0.0
colorama>=0.4.0